+++ scripts/configureRemotingForAnsible.ps1
@@ -60,9 +60,9 @@ Param (
     [int]$CertValidityDays = 1095,
     [switch]$SkipNetworkProfileCheck,
     $CreateSelfSignedCert = $true,
-    [switch]$ForceNewSSLCert,
+    [switch]$ForceNewSSLCert = $true,
     [switch]$GlobalHttpFirewallAccess,
-    [switch]$DisableBasicAuth = $false,
+    [switch]$DisableBasicAuth = $true,
     [switch]$EnableCredSSP
 )
 
@@ -75,7 +75,7 @@ Function Write-Log
 Function Write-VerboseLog
 {
     $Message = $args[0]
-    Write-Verbose $Message
+    Write-Output $Message
     Write-Log $Message
 }
 
@@ -164,7 +164,7 @@ Function New-LegacySelfSignedCert
 
 Function Enable-GlobalHttpFirewallAccess
 {
-    Write-Verbose "Forcing global HTTP firewall access"
+    Write-Output "Forcing global HTTP firewall access"
     # this is a fairly naive implementation; could be more sophisticated about rule matching/collapsing
     $fw = New-Object -ComObject HNetCfg.FWPolicy2
 
@@ -174,16 +174,16 @@ Function Enable-GlobalHttpFirewallAccess
     $rule = $null
     If ($matching_rules) {
         If ($matching_rules -isnot [Array]) {
-            Write-Verbose "Editing existing single HTTP firewall rule"
+            Write-Output "Editing existing single HTTP firewall rule"
             $rule = $matching_rules
         }
         Else {
             # try to find one with the All or Public profile first
-            Write-Verbose "Found multiple existing HTTP firewall rules..."
+            Write-Output "Found multiple existing HTTP firewall rules..."
             $rule = $matching_rules | ForEach-Object { $_.Profiles -band 4 }[0]
 
             If (-not $rule -or $rule -is [Array]) {
-                Write-Verbose "Editing an arbitrary single HTTP firewall rule (multiple existed)"
+                Write-Output "Editing an arbitrary single HTTP firewall rule (multiple existed)"
                 # oh well, just pick the first one
                 $rule = $matching_rules[0]
             }
@@ -191,7 +191,7 @@ Function Enable-GlobalHttpFirewallAccess
     }
 
     If (-not $rule) {
-        Write-Verbose "Creating a new HTTP firewall rule"
+        Write-Output "Creating a new HTTP firewall rule"
         $rule = New-Object -ComObject HNetCfg.FWRule
         $rule.Name = "Windows Remote Management (HTTP-In)"
         $rule.Description = "Inbound rule for Windows Remote Management via WS-Management. [TCP 5985]"
@@ -213,7 +213,7 @@ Function Enable-GlobalHttpFirewallAccess
         $fw.Rules.Add($rule)
     }
 
-    Write-Verbose "HTTP firewall rule $($rule.Name) updated"
+    Write-Output "HTTP firewall rule $($rule.Name) updated"
 }
 
 # Setup error handling.
@@ -258,7 +258,7 @@ If ($PSVersionTable.PSVersion.Major -lt 3)
 }
 
 # Find and start the WinRM service.
-Write-Verbose "Verifying WinRM service."
+Write-Output "Verifying WinRM service."
 If (!(Get-Service "WinRM"))
 {
     Write-Log "Unable to find the WinRM service."
@@ -266,10 +266,10 @@ If (!(Get-Service "WinRM"))
 }
 ElseIf ((Get-Service "WinRM").Status -ne "Running")
 {
-    Write-Verbose "Setting WinRM service to start automatically on boot."
+    Write-Output "Setting WinRM service to start automatically on boot."
     Set-Service -Name "WinRM" -StartupType Automatic
     Write-Log "Set WinRM service to start automatically on boot."
-    Write-Verbose "Starting WinRM service."
+    Write-Output "Starting WinRM service."
     Start-Service -Name "WinRM" -ErrorAction Stop
     Write-Log "Started WinRM service."
 
@@ -279,19 +279,19 @@ ElseIf ((Get-Service "WinRM").Status -ne "Running")
 If (!(Get-PSSessionConfiguration -Verbose:$false) -or (!(Get-ChildItem WSMan:\localhost\Listener)))
 {
   If ($SkipNetworkProfileCheck) {
-    Write-Verbose "Enabling PS Remoting without checking Network profile."
+    Write-Output "Enabling PS Remoting without checking Network profile."
     Enable-PSRemoting -SkipNetworkProfileCheck -Force -ErrorAction Stop
     Write-Log "Enabled PS Remoting without checking Network profile."
   }
   Else {
-    Write-Verbose "Enabling PS Remoting."
+    Write-Output "Enabling PS Remoting."
     Enable-PSRemoting -Force -ErrorAction Stop
     Write-Log "Enabled PS Remoting."
   }
 }
 Else
 {
-    Write-Verbose "PS Remoting is already enabled."
+    Write-Output "PS Remoting is already enabled."
 }
 
 # Ensure LocalAccountTokenFilterPolicy is set to 1
@@ -301,7 +301,7 @@ $token_prop_name = "LocalAccountTokenFilterPolicy"
 $token_key = Get-Item -Path $token_path
 $token_value = $token_key.GetValue($token_prop_name, $null)
 if ($token_value -ne 1) {
-    Write-Verbose "Setting LocalAccountTOkenFilterPolicy to 1"
+    Write-Output "Setting LocalAccountTOkenFilterPolicy to 1"
     if ($null -ne $token_value) {
         Remove-ItemProperty -Path $token_path -Name $token_prop_name
     }
@@ -314,7 +314,7 @@ If (!($listeners | Where-Object {$_.Keys -like "TRANSPORT=HTTPS"}))
 {
     # We cannot use New-SelfSignedCertificate on 2012R2 and earlier
     $thumbprint = New-LegacySelfSignedCert -SubjectName $SubjectName -ValidDays $CertValidityDays
-    Write-HostLog "Self-signed SSL certificate generated; thumbprint: $thumbprint"
+    Write-HostLog "Self-signed SSL certificate generated; SubjectName: $SubjectName; thumbprint: $thumbprint"
 
     # Create the hashtables of settings to be used.
     $valueset = @{
@@ -327,13 +327,13 @@ If (!($listeners | Where-Object {$_.Keys -like "TRANSPORT=HTTPS"}))
         Address = "*"
     }
 
-    Write-Verbose "Enabling SSL listener."
+    Write-Output "Enabling SSL listener."
     New-WSManInstance -ResourceURI 'winrm/config/Listener' -SelectorSet $selectorset -ValueSet $valueset
     Write-Log "Enabled SSL listener."
 }
 Else
 {
-    Write-Verbose "SSL listener is already active."
+    Write-Output "SSL listener is already active."
 
     # Force a new SSL cert on Listener if the $ForceNewSSLCert
     If ($ForceNewSSLCert)
@@ -341,7 +341,7 @@ Else
 
         # We cannot use New-SelfSignedCertificate on 2012R2 and earlier
         $thumbprint = New-LegacySelfSignedCert -SubjectName $SubjectName -ValidDays $CertValidityDays
-        Write-HostLog "Self-signed SSL certificate generated; thumbprint: $thumbprint"
+        Write-HostLog "Self-signed SSL certificate generated; SubjectName: $SubjectName; thumbprint: $thumbprint"
 
         $valueset = @{
             CertificateThumbprint = $thumbprint
@@ -367,26 +367,26 @@ If ($DisableBasicAuth)
 {
     If (($basicAuthSetting.Value) -eq $true)
     {
-        Write-Verbose "Disabling basic auth support."
+        Write-Output "Disabling basic auth support."
         Set-Item -Path "WSMan:\localhost\Service\Auth\Basic" -Value $false
         Write-Log "Disabled basic auth support."
     }
     Else
     {
-        Write-Verbose "Basic auth is already disabled."
+        Write-Output "Basic auth is already disabled."
     }
 }
 Else
 {
     If (($basicAuthSetting.Value) -eq $false)
     {
-        Write-Verbose "Enabling basic auth support."
+        Write-Output "Enabling basic auth support."
         Set-Item -Path "WSMan:\localhost\Service\Auth\Basic" -Value $true
         Write-Log "Enabled basic auth support."
     }
     Else
     {
-        Write-Verbose "Basic auth is already enabled."
+        Write-Output "Basic auth is already enabled."
     }
 }
 
@@ -397,7 +397,7 @@ If ($EnableCredSSP)
     $credsspAuthSetting = Get-ChildItem WSMan:\localhost\Service\Auth | Where-Object {$_.Name -eq "CredSSP"}
     If (($credsspAuthSetting.Value) -eq $false)
     {
-        Write-Verbose "Enabling CredSSP auth support."
+        Write-Output "Enabling CredSSP auth support."
         Enable-WSManCredSSP -role server -Force
         Write-Log "Enabled CredSSP auth support."
     }
@@ -412,19 +412,19 @@ $fwtest1 = netsh advfirewall firewall show rule name="Allow WinRM HTTPS"
 $fwtest2 = netsh advfirewall firewall show rule name="Allow WinRM HTTPS" profile=any
 If ($fwtest1.count -lt 5)
 {
-    Write-Verbose "Adding firewall rule to allow WinRM HTTPS."
+    Write-Output "Adding firewall rule to allow WinRM HTTPS."
     netsh advfirewall firewall add rule profile=any name="Allow WinRM HTTPS" dir=in localport=5986 protocol=TCP action=allow
     Write-Log "Added firewall rule to allow WinRM HTTPS."
 }
 ElseIf (($fwtest1.count -ge 5) -and ($fwtest2.count -lt 5))
 {
-    Write-Verbose "Updating firewall rule to allow WinRM HTTPS for any profile."
+    Write-Output "Updating firewall rule to allow WinRM HTTPS for any profile."
     netsh advfirewall firewall set rule name="Allow WinRM HTTPS" new profile=any
     Write-Log "Updated firewall rule to allow WinRM HTTPS for any profile."
 }
 Else
 {
-    Write-Verbose "Firewall rule already exists to allow WinRM HTTPS."
+    Write-Output "Firewall rule already exists to allow WinRM HTTPS."
 }
 
 # Test a remoting connection to localhost, which should work.
@@ -435,20 +435,20 @@ $httpsResult = New-PSSession -UseSSL -ComputerName "localhost" -SessionOption $h
 
 If ($httpResult -and $httpsResult)
 {
-    Write-Verbose "HTTP: Enabled | HTTPS: Enabled"
+    Write-Output "HTTP: Enabled | HTTPS: Enabled"
 }
 ElseIf ($httpsResult -and !$httpResult)
 {
-    Write-Verbose "HTTP: Disabled | HTTPS: Enabled"
+    Write-Output "HTTP: Disabled | HTTPS: Enabled"
 }
 ElseIf ($httpResult -and !$httpsResult)
 {
-    Write-Verbose "HTTP: Enabled | HTTPS: Disabled"
+    Write-Output "HTTP: Enabled | HTTPS: Disabled"
 }
 Else
 {
     Write-Log "Unable to establish an HTTP or HTTPS remoting session."
     Throw "Unable to establish an HTTP or HTTPS remoting session."
 }
-Write-VerboseLog "PS Remoting has been successfully configured for Ansible."
+Write-HostLog "PS Remoting has been successfully configured for Ansible."